# Cherry-pick #16849 to 7.x: [Winlogbeat] Add experimental event log reader with increased performance