--- a/winlogbeat/sys/wineventlog/template.go
+++ b/winlogbeat/sys/wineventlog/template.go
@@ -15,12 +15,21 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// Package wineventlog provides access to the Windows Event Log API used in
-// all versions of Windows since Vista (i.e. Windows 7+  and Windows Server 2008+).
-// This is distinct from the Event Logging API that was used in Windows XP,
-// Windows Server 2003, and Windows 2000.
 package wineventlog
 
-// Add -trace to enable debug prints around syscalls.
-//go:generate go get golang.org/x/sys/windows/mkwinsyscall
-//go:generate $GOPATH/bin/mkwinsyscall.exe -systemdll -output zsyscall_windows.go syscall_windows.go
+import (
+	"encoding/xml"
+)
+
+type eventTemplate struct {
+	Data []eventData `xml:"data"`
+}
+
+type eventData struct {
+	Name string `xml:"name,attr"`
+	Type string `xml:"outType,attr"`
+}
+
+func (t *eventTemplate) Unmarshal(xmlData []byte) error {
+	return xml.Unmarshal(xmlData, t)
+}