--- a/winlogbeat/sys/wineventlog/wineventlog_windows_test.go
+++ b/winlogbeat/sys/wineventlog/wineventlog_windows_test.go
@@ -108,7 +108,3 @@ func TestChannels(t *testing.T) {
 		}
 	}
 }
-
-func TestExtension(t *testing.T) {
-	filepath.Ext("sysmon")
-}