… winlogbeat/eventlog/bench_test.go
… winlogbeat/eventlog/eventlogging_test.go
… winlogbeat/eventlog/wineventlog_test.go
… winlogbeat/sys/wineventlog/bookmark_test.go
… winlogbeat/sys/wineventlog/format_message_test.go
… winlogbeat/sys/wineventlog/iterator.go
… winlogbeat/sys/wineventlog/iterator_test.go
… winlogbeat/sys/wineventlog/metadata_store.go
… winlogbeat/sys/wineventlog/metadata_store_test.go
… winlogbeat/sys/wineventlog/publisher_metadata.go
… winlogbeat/sys/wineventlog/publisher_metadata_test.go
… winlogbeat/sys/wineventlog/renderer.go
… winlogbeat/sys/wineventlog/renderer_test.go
… winlogbeat/sys/wineventlog/stringinserts_test.go
… winlogbeat/sys/wineventlog/syscall_windows.go
… winlogbeat/sys/wineventlog/template_test.go
… winlogbeat/sys/wineventlog/testdata/application-windows-error-reporting.evtx
… winlogbeat/sys/wineventlog/util_test.go
… winlogbeat/sys/wineventlog/winmeta.go
… winlogbeat/sys/wineventlog/zsyscall_windows.go